get rid of primitive enums

+++ migrations/2019-02-09-212559_orders/up.sql
@@ -4,7 +4,7 @@ CREATE TABLE orders (
   merchant_id TEXT NOT NULL,
   grin_amount BIGINT NOT NULL,
   amount JSONB NOT NULL,
-  status INTEGER NOT NULL,
+  status SMALLINT NOT NULL,
   confirmations INTEGER NOT NULL,
   email TEXT,
   created_at TIMESTAMP NOT NULL,
